+++ playground/src/components/nav.vue
@@ -1,19 +1,18 @@
 <script lang="ts" setup>
 import { CSSProperties } from "@toolmain/shared"
 import { Resize } from "@toolmain/components"
-import { ref, useTemplateRef } from "vue"
-const scaleRef = useTemplateRef("scale")
+import { ref } from "vue"
 const targetStyle = ref<CSSProperties>({
   width: "350px",
 })
 </script>
 <template>
   <div class="subnav-container">
-    <div class="subnav-provider" :style="targetStyle" ref="scale">
+    <div class="subnav-provider" :style="targetStyle">
+      <Resize v-model="targetStyle" size="8px" direction="right" />
       <el-card class="subnav-card" body-class="flex flex-1 flex-col overflow-hidden" shadow="never">
         <slot name="submenu"></slot>
       </el-card>
-      <Resize v-model="targetStyle" size="8px" direction="right" :target="scaleRef" />
     </div>
     <div class="subnav-content">
       <slot name="content"></slot>